Moved all functionality to runtime loader
