+++ src/software/thunderscope/constants.py
@@ -383,8 +383,8 @@ class AISelectorConstants:
 
     AI_FOLDER = "/opt/tbotspython/external_ai"
     CONFIG_FILE_PATH = AI_FOLDER + "/ai_config.toml"
-    FULLSYSTEM_DISPLAY_NAME = "Default Fullsystem" #TODO is there a better name?
-    FULLSYSTEM_PATH = "software/unix_full_system" #TODO is this right? Should I use absolute path?
+    FULLSYSTEM_DISPLAY_NAME = "Default Fullsystem"
+    FULLSYSTEM_PATH = "software/unix_full_system"
     BLUE_TOML_FIELD = "blue_path_to_binary"
     YELLOW_TOML_FIELD = "yellow_path_to_binary"
